+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+This version has been generated using:
+
+- [Connector code generator](https://github.com/datafoodconsortium/connector-codegen) [main](https://github.com/datafoodconsortium/connector-codegen/commit/e630d0334b70e4a7e7e07a9151fecaf53af0557e).
+- [UML data model](https://github.com/datafoodconsortium/data-model-uml) [next](https://github.com/datafoodconsortium/data-model-uml/commit/621e823c21c79a58b117bae97132da9140e47be6).
+
+### Added
+
+- Add the `soldBy` property in `Order`.
+- In `DefinedProduct`:
+  - Add the `hasVariant` property.
+  - Add the `isVariantOf` property.
+
 ## [1.0.0-alpha.13] - 2024-10-11
 
 This version has been generated using: